Updated the code for Python 3.
The current vehicle signal specification utility was created in Python 2 and
has remained a Python 2 utility ever since.  There is a need to be able to
support Python 3 with this utility so an update of the Python code files is
necessary.

The Python "2to3" conversion utility has been run on all of the Python source
files to modify them for use with Python 3.

Note: This should probably also include a change in the protocol version
number.  The current version code is "2.0.0-alpha+004".  Please specify how
this should be updated for this code change.