Improving documentation for TravelledDistance
Previously Vehicle.TravelledDistance and Transmission.TravelledDistance
had the same description. In VSS meeting it was discussed whether they
could be considered as duplicates or if they actually have different
meaning. It was agreed that they might be different, if the transmission
has been replaced you might want to keep track of the wear/usage of the
new/replacement transmission separately.